# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminToken {
  token: string;
}

export interface WorksVideo {
  id: number;
  title: string;
  /** music-video | bts | reels */
  videoType: string;
  /**
     * Object storage path for uploaded video files
     * @nullable
     */
  storagePath?: string | null;
  /**
     * YouTube video ID for embedded videos
     * @nullable
     */
  youtubeId?: string | null;
  /**
     * Object storage path for thumbnail image
     * @nullable
     */
  thumbnailPath?: string | null;
  position: number;
  createdAt: string;
}

export interface WorksVideoInput {
  /** @minLength 1 */
  title: string;
  videoType: string;
  /** @nullable */
  storagePath?: string | null;
  /** @nullable */
  youtubeId?: string | null;
  /** @nullable */
  thumbnailPath?: string | null;
  position?: number;
}

export interface WorksVideoUpdate {
  title?: string;
  videoType?: string;
  /** @nullable */
  storagePath?: string | null;
  /** @nullable */
  youtubeId?: string | null;
  /** @nullable */
  thumbnailPath?: string | null;
  position?: number;
}

export interface GalleryImage {
  id: number;
  /** @nullable */
  altText?: string | null;
  storagePath: string;
  position: number;
  createdAt: string;
}

export interface GalleryImageInput {
  storagePath: string;
  /** @nullable */
  altText?: string | null;
  position?: number;
}

export interface GalleryImageUpdate {
  /** @nullable */
  altText?: string | null;
  position?: number;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}