# lib/api-zod/src/generated/types/galleryImage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GalleryImage {
  id: number;
  /** @nullable */
  altText?: string | null;
  storagePath: string;
  position: number;
  createdAt: string;
}