# lib/api-zod/src/generated/types/worksVideo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorksVideo {
  id: number;
  title: string;
  /** music-video | bts | reels */
  videoType: string;
  /**
     * Object storage path for uploaded video files
     * @nullable
     */
  storagePath?: string | null;
  /**
     * YouTube video ID for embedded videos
     * @nullable
     */
  youtubeId?: string | null;
  /**
     * Object storage path for thumbnail image
     * @nullable
     */
  thumbnailPath?: string | null;
  position: number;
  createdAt: string;
}