# lib/api-zod/src/generated/types/worksVideoUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorksVideoUpdate {
  title?: string;
  videoType?: string;
  /** @nullable */
  storagePath?: string | null;
  /** @nullable */
  youtubeId?: string | null;
  /** @nullable */
  thumbnailPath?: string | null;
  position?: number;
}